backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
cryptography==41.0.3
opentelemetry-sdk==1.19.0
opentelemetry-api==1.19.0
opentelemetry-exporter-gcp-trace==1.5.0
opentelemetry-propagator-gcp==1.5.0
opentelemetry-instrumentation-fastapi==0.40b0
email-validator==2.0
google-cloud-firestore==2.11.1